KAS-1137 refactor out old code and isRemark logic
