+++ util/index.js
@@ -1,0 +1,17 @@
+
+const parseSparqlResults = (data) => {
+  const vars = data.head.vars;
+  return data.results.bindings.map(binding => {
+    let obj = {};
+    vars.forEach(varKey => {
+      if (binding[varKey]) {
+        obj[varKey] = binding[varKey].value;
+      }
+    });
+    return obj;
+  })
+};
+
+export {
+  parseSparqlResults
+};
